Limit test utils to a single environment (browser / ssr).
Refactors various test utils so that tests are run for a requested browser (or ssr) rather than for the whole permuatation of browsers and React versions.
Travis CI is replaced with GitHub workflows. Jobs for various browsers are run in parallel rather than sequentially.
BrowserStack connectivity issues are mitigated with BS-specific settings for Karma. Karma launcher for BS is set to v1.4 (more stable than newer versions).